# Conflicts:
#	src/main/java/com/sparta/second/controller/UserController.java
#	src/main/java/com/sparta/second/service/UserService.java

+++ src/main/java/com/sparta/second/service/ReviewService.java
@@ -4,8 +4,10 @@ import com.sparta.second.dto.ReviewListResponseDto;
 import com.sparta.second.dto.ReviewRequestDto;
 import com.sparta.second.dto.ReviewResponseDto;
 import com.sparta.second.entity.Review;
+import com.sparta.second.entity.Shop;
 import com.sparta.second.entity.User;
 import com.sparta.second.repository.ReviewRepository;
+import com.sparta.second.repository.ShopRepository;
 import com.sparta.second.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,11 +21,11 @@ import java.util.stream.Collectors;
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
-//    private final ShopRepository shopRepository;
+    private final ShopRepository shopRepository;
     public void createReview(UserDetailsImpl userDetails, ReviewRequestDto reviewRequestDto) {
-//        Shop shop = shopRepository.findById(reviewRequestDto.getShopId()).orElseThrow(() -> new RuntimeException());
+        Shop shop = shopRepository.findById(reviewRequestDto.getShopId()).orElseThrow(() -> new RuntimeException());
 
-        Review review = new Review(userDetails.getUser(),reviewRequestDto);
+        Review review = new Review(userDetails.getUser(),reviewRequestDto,shop);
 
         reviewRepository.save(review);
     }
@@ -46,9 +48,9 @@ public class ReviewService {
     public void update(User user, Long reviewId, ReviewRequestDto reviewRequestDto) {
         Review review = findReview(reviewId);
 
-//        if (!review.getUser().equals(user)) {
-//            throw new RejectedExecutionException();
-//        }
+        if (!review.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
         review.update(reviewRequestDto);
 
         reviewRepository.save(review);
@@ -58,9 +60,9 @@ public class ReviewService {
         // 해당 포스트가 DB에 존재하는지 확인
         Review review = findReview(reviewId);
         //리뷰 작성자와 요청자(user)가 같은지 학인
-//        if (!review.getUser().equals(user)) {
-//            throw new RejectedExecutionException();
-//        }
+        if (!review.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
         // post 삭제
         reviewRepository.delete(review);
     }
